sighook/requirements.txt
about-time==4.2.1
aiodns==3.1.1
aiohttp==3.9.1
aiosignal==1.3.1
alive-progress==3.1.5
async-timeout==4.0.3
attrs==23.1.0
Bottleneck==1.3.5
ccxt==4.1.74
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
Flask==2.2.2
frozenlist==1.4.0
grapheme==0.6.0
idna==3.4
itsdangerous==2.0.1
Jinja2==3.1.2
MarkupSafe==2.1.1
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
multidict==6.0.4
numexpr==2.8.7
numpy
pandas==2.1.4
pycares==4.4.0
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
Werkzeug==2.3.6
yarl==1.9.3
